Bug 1551735 - Split out screenshot grabbing infrasturcture into a new module r=kvark
Differential Revision: https://phabricator.services.mozilla.com/D32523

[wrupdater] From https://hg.mozilla.org/mozilla-central/rev/a7868d694fe115c432015a01be88cf44d88308d9
