+++ webrender/src/lib.rs
@@ -115,6 +115,7 @@ mod renderer;
 mod resource_cache;
 mod scene;
 mod scene_builder;
+mod screen_capture;
 mod segment;
 mod shade;
 mod spatial_node;
@@ -212,10 +213,11 @@ pub use crate::device::{ProgramBinary, ProgramCache, ProgramCacheObserver};
 pub use crate::device::Device;
 pub use crate::frame_builder::ChasePrimitive;
 pub use crate::profiler::{ProfilerHooks, set_profiler_hooks};
-pub use crate::renderer::{AsyncPropertySampler, AsyncScreenshotHandle, CpuProfile, DebugFlags};
+pub use crate::renderer::{AsyncPropertySampler, CpuProfile, DebugFlags};
 pub use crate::renderer::{OutputImageHandler, RendererKind, ExternalImage, ExternalImageHandler};
 pub use crate::renderer::{ExternalImageSource, GpuProfile, GraphicsApi, GraphicsApiInfo, PipelineInfo};
 pub use crate::renderer::{Renderer, RendererOptions, RenderResults, RendererStats, SceneBuilderHooks};
 pub use crate::renderer::{ThreadListener, ShaderPrecacheFlags, MAX_VERTEX_TEXTURE_WIDTH};
+pub use crate::screen_capture::AsyncScreenshotHandle;
 pub use crate::shade::{Shaders, WrShaders};
 pub use api as webrender_api;
